backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
pydantic==2.11.1
pydantic-settings==2.7.1
email-validator==2.2.0
firebase-admin==6.6.0
google-cloud-firestore==2.20.0
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20
